Adding address model and link it to customer model
It was added more two models beyond <code>address</code> model: <code>country</code> and <code>state</code>. Those models are very similar, the attributes are the same (<code>name</code> and <code>abbreviation</code>), for example. Using this, it's possible to define, in the future, how many customers are from one defined state, country, city etc. None of the <code>address</code> attributes are required and was created a partial form to this model that could be included in the form of each model that has a relationship with it. Fixes #37. :smile: