Updating rails version
When I was solving an issue to install and configure the postgresql in my mac I tried up to date the rails version to 'solve' this issue. I solved it, but the problem was not with rails version. So, I ran all tests (everything is green :white_check_mark:) and decided that we could use this version of rails. :wink: :satisfied: